backend/requirements.txt
fastapi
slowapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
cloudinary
debugpy